HuggingFaceDeployment/requirements.txt
langchain
langchain_openai
openai
huggingface_hub
python-dotenv
streamlit